+++ .gitignore
@@ -1,0 +1,3 @@
+__py_cache__/
+venv/
+.pytest_chace
